requirements.txt
PyYAML==3.12
pylint==1.8.2
pep8==1.7.1
autopep8==1.3.4
requests
